added  struct, changed decoding and predictor API to allow in-place modification

+++ src/tiff/error.rs
@@ -159,7 +159,7 @@ pub enum TiffUnsupportedError {
     UnsupportedSampleDepth(u8),
     UnsupportedSampleFormat(Vec<SampleFormat>),
     // UnsupportedColorType(ColorType),
-    UnsupportedBitsPerChannel(u8),
+    UnsupportedBitsPerChannel(u16),
     UnsupportedPlanarConfig(Option<PlanarConfiguration>),
     UnsupportedDataType,
     UnsupportedInterpretation(PhotometricInterpretation),
